requirements.txt
PySide6>=6.5.0
cryptography>=41.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
